fix unsigned comparison

+++ rclcpp/test/test_time.cpp
@@ -66,7 +66,7 @@ TEST(TestTime, convertions) {
   EXPECT_EQ(
     RCL_S_TO_NS(static_cast<uint64_t>(msg.sec)) + static_cast<uint64_t>(msg.nanosec),
     time.nanoseconds());
-  EXPECT_EQ(msg.sec, RCL_NS_TO_S(time.nanoseconds()));
+  EXPECT_EQ(static_cast<uint64_t>(msg.sec), RCL_NS_TO_S(time.nanoseconds()));
 
   builtin_interfaces::msg::Time negative_time_msg;
   negative_time_msg.sec = -1;
